refactor: function abstraction within compliances, colors coherence
