+++ ui/lib/compliance/aws-well-architected.tsx
@@ -8,26 +8,24 @@ import {
   AWSWellArchitectedAttributesMetadata,
   Framework,
   Requirement,
-  RequirementItemData,
   RequirementsData,
   RequirementStatus,
 } from "@/types/compliance";
 
-import { updateCounters } from "./commons";
+import {
+  calculateFrameworkCounters,
+  createRequirementsMap,
+  findOrCreateCategory,
+  findOrCreateControl,
+  findOrCreateFramework,
+} from "./commons";
 
 export const mapComplianceData = (
   attributesData: AttributesData,
   requirementsData: RequirementsData,
 ): Framework[] => {
   const attributes = attributesData?.data || [];
-  const requirements = requirementsData?.data || [];
-
-  // Create a map for quick lookup of requirements by id
-  const requirementsMap = new Map<string, RequirementItemData>();
-  requirements.forEach((req: RequirementItemData) => {
-    requirementsMap.set(req.id, req);
-  });
-
+  const requirementsMap = createRequirementsMap(requirementsData);
   const frameworks: Framework[] = [];
 
   // Process attributes and merge with requirements data
@@ -54,44 +52,14 @@ export const mapComplianceData = (
       continue;
     }
 
-    // Find or create framework
-    let framework = frameworks.find((f) => f.name === frameworkName);
-    if (!framework) {
-      framework = {
-        name: frameworkName,
-        pass: 0,
-        fail: 0,
-        manual: 0,
-        categories: [],
-      };
-      frameworks.push(framework);
-    }
+    // Find or create framework using common helper
+    const framework = findOrCreateFramework(frameworks, frameworkName);
 
-    // Find or create category (Section)
-    let category = framework.categories.find((c) => c.name === sectionName);
-    if (!category) {
-      category = {
-        name: sectionName,
-        pass: 0,
-        fail: 0,
-        manual: 0,
-        controls: [],
-      };
-      framework.categories.push(category);
-    }
+    // Find or create category (Section) using common helper
+    const category = findOrCreateCategory(framework.categories, sectionName);
 
-    // Find or create control (SubSection)
-    let control = category.controls.find((c) => c.label === subSectionName);
-    if (!control) {
-      control = {
-        label: subSectionName,
-        pass: 0,
-        fail: 0,
-        manual: 0,
-        requirements: [],
-      };
-      category.controls.push(control);
-    }
+    // Find or create control (SubSection) using common helper
+    const control = findOrCreateControl(category.controls, subSectionName);
 
     // Create requirement
     const finalStatus: RequirementStatus = status as RequirementStatus;
@@ -114,36 +82,8 @@ export const mapComplianceData = (
     control.requirements.push(requirement);
   }
 
-  // Calculate counters
-  frameworks.forEach((framework) => {
-    framework.pass = 0;
-    framework.fail = 0;
-    framework.manual = 0;
-
-    framework.categories.forEach((category) => {
-      category.pass = 0;
-      category.fail = 0;
-      category.manual = 0;
-
-      category.controls.forEach((control) => {
-        control.pass = 0;
-        control.fail = 0;
-        control.manual = 0;
-
-        control.requirements.forEach((requirement) => {
-          updateCounters(control, requirement.status);
-        });
-
-        category.pass += control.pass;
-        category.fail += control.fail;
-        category.manual += control.manual;
-      });
-
-      framework.pass += category.pass;
-      framework.fail += category.fail;
-      framework.manual += category.manual;
-    });
-  });
+  // Calculate counters using common helper
+  calculateFrameworkCounters(frameworks);
 
   return frameworks;
 };
